server/requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
python-dotenv==1.0.1
openai==1.41.0
pymongo==4.8.0
pydantic==2.7.4